% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_null_glm_SCANG.r
\name{fit_null_glm_SCANG}
\alias{fit_null_glm_SCANG}
\title{Fit generalized linear model under the null hypothesis for unrelated samples.}
\usage{
fit_null_glm_SCANG(fixed, data, family = binomial(link = "logit"),
  times = 2000, ...)
}
\arguments{
\item{fixed}{an object of class \code{\link{formula}} (or one that can be coerced to that class):
a symbolic description of the fixed effects model to be fitted.}

\item{data}{a data frame or list (or object coercible by \code{\link{as.data.frame}} to a data frame)
containing the variables in the model.}

\item{family}{a description of the error distribution and link function to be used
in the model. This can be a character string naming a family function, a family
function or the result of a call to a family function. (See \code{\link{family}} for details of family functions).
Can be either "gaussian" for continuous phenotype or "binomial" for binary phenotype.}

\item{times}{a number of pseudo-residuals (default = 2000).}

\item{...}{additional arguments that could be passed to \code{\link{glm}}.}
}
\value{
The function returns an object of the model fit from \code{\link{glm}} (\code{obj_nullmodel}),
with an additional element indicating the samples are unrelated
(\code{obj_nullmodel$relatedness = FALSE}). See \code{\link{glm}} for more details.
}
\description{
The \code{fit_null_glm_SCANG} function is a wrapper of the \code{\link{glm}} function from the
\code{\link{stats}} package that fits a regression model under the null hypothesis for
unrelated samples, which provides the preliminary step for subsequent
variant-set tests in whole
genome sequencing data analysis.
}
