% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCANG.r
\name{SCANG}
\alias{SCANG}
\title{SCANG procedure using omnibus test}
\usage{
SCANG(genotype, obj_nullmodel, Lmin, Lmax, annotation_phred = NULL,
  rare_maf_cutoff = 0.05, steplength = 5, alpha = 0.05,
  filter = 1e-04, f = 0.5, subseq_num = 2000)
}
\arguments{
\item{genotype}{an n*p genotype matrix (dosage matrix) of the target sequence, where n is the sample
size and p is the number of variants.}

\item{obj_nullmodel}{an object from fitting the null model, which is the
output from either \code{\link{fit_null_glm_SCANG}} function for unrelated samples or
\code{\link{fit_null_glmmkin_SCANG}} function for related samples. Note that \code{\link{fit_null_glmmkin_SCANG}}
is a wrapper of \code{\link{glmmkin}} function from the \code{\link{GMMAT}} package.}

\item{Lmin}{minimum number of variants in searching windows.}

\item{Lmax}{maximum number of variants in searching windows.}

\item{annotation_phred}{a data frame or matrix of functional annotation data
of dimension p*q (or a vector of a single annotation score with length p).
Continuous scores should be given in PHRED score scale, where the PHRED score
of j-th variant is defined to be -10*log10(rank(-score_j)/total) across the genome. (Binary)
categorical scores should be taking values 0 or 1, where 1 is functional and 0 is
non-functional. If not provided, SCANG will perform the original procedure without annotations (default = NULL).}

\item{rare_maf_cutoff}{the cutoff of maximum minor allele frequency in
defining rare variants (default = 0.05).}

\item{steplength}{difference of number of variants in searching windows, that is, the number of variants in
searching windows are Lmin, Lmin+steplength, Lmin+steplength,...,Lmax (default = 5).}

\item{alpha}{familty-wise/genome-wide significance level (default = 0.05).}

\item{filter}{a filtering threshold of screening method for SKAT. SKAT p-values are calculated for regions whose p-value
is possibly smaller than the filtering threshold (default = 1e-4).}

\item{f}{an overlap fraction, which controls for the overlapping proportion of of detected regions. For example,
when f=0, the detected regions are non-overlapped with each other,
and when f=1, we keep every susceptive region as detected regions (default = 0.5).}

\item{subseq_num}{number of variants run in each sub-sequence (default = 2000).}
}
\value{
The function returns a list with the following members:

\code{SCANG_O_res}:   A matrix that summarized the significant region detected by SCANG-O.
The first column is the -log(p-value) of the detected region.
The next two columns are the location of the detected region (in sense of variants order).
The last column is the family-wise/genome-wide error rate of the detected region.
The result (0,0,0,1) means there is no significant region.

\code{SCANG_O_top1}:  A vector of length 4 which summarized the top 1 region detected by SCANG-O.
The first element is the -log(p-value) of the region.
The next two elements are the location of the detected region (in sense of variants order).
The last element is the family-wise/genome-wide p-value.

\code{SCANG_O_thres}: Empirical threshold of SCANG-O for controlling the family-wise type I error at alpha level.

\code{SCANG_O_thres_boot}: A vector of Monte Carlo simulation sample for generating the empirical threshold. The 1-alpha quantile of this vector is
the empirical threshold.

\code{SCANG_S_res, SCANG_S_thres, SCANG_S_top1, SCANG_S_thres_boot}: Analysis
results using SCANG-S. Details see SCANG-O.

\code{SCANG_B_res, SCANG_B_thres, SCANG_B_top1, SCANG_B_thres_boot}: Analysis
results using SCANG-B. Details see SCANG-O.
}
\description{
The \code{SCANG} function takes in genotype and the object from fitting the null
model and detect the association between a
quantitative/dichotomous phenotype and a variant-set in a sequence by using SCANG procedure, including SCANG-O, SCANG-B and SCANG-S.
For each region, the scan statistic of SCANG-O is the set-based p-value of STAAR-O, which is an omnibus test that aggregated p-values
across different types of multiple annotation-weighted variant-set tests SKAT(1,1), SKAT(1,25), Burden(1,1) and Burden(1,25) using ACAT method;
the scan statistic of SCANG-S is the set-based p-value of STAAR-S, which is an omnibus test that aggregated p-values
across multiple annotation-weighted variant-set tests SKAT(1,1) and SKAT(1,25) using ACAT method;
the scan statistic of SCANG-B is the set-based p-value of STAAR-B, which is an omnibus test that aggregated p-values
across multiple annotation-weighted variant-set tests Burden(1,1) and Burden(1,25) using ACAT method.
}
\references{
Li, Z., et al. (2019). Dynamic Scan Procedure for Detecting Rare-Variant
Association Regions in Whole-Genome Sequencing Studies.
\emph{The American Journal of Human Genetics 104}(5), 802-814.
(\href{https://pubmed.ncbi.nlm.nih.gov/30982610/}{pub})

Li, X., Li, Z., et al. (2020). Dynamic incorporation of multiple
in-silico functional annotations empowers rare variant association analysis of
large whole genome sequencing studies at scale.
\emph{Nature Genetics (in press)}.

Liu, Y., et al. (2019). Acat: A fast and powerful p value combination
method for rare-variant analysis in sequencing studies.
\emph{The American Journal of Human Genetics 104}(3), 410-421.
(\href{https://www.sciencedirect.com/science/article/pii/S0002929719300023}{pub})
}
